Renames network.d.ts to networks.d.ts
